+++ packages/lumx-react/src/stories/utils/CustomLink.tsx
@@ -1,0 +1,7 @@
+/* eslint-disable jsx-a11y/anchor-has-content,jsx-a11y/anchor-is-valid */
+import React from 'react';
+
+/**
+ * Example custom link that can be used in linkAs props.
+ */
+export const CustomLink: any = ({ ...props }) => <a href="#" data-custom-link {...props} />;